Fixed 'size' variable size to unsigned long in loadfile().

+++ src/fdhandler/fdhandler.c
@@ -48,7 +48,7 @@ DATA *loadfile(char *path)
 	//Win32API
 #else
 	int fd; /* file descriptor */
-	int size; /* file size */
+	unsigned long size; /* file size */
 	int bytesread; /* bytes read from file */
 	DATA *data; /* the structure to be returned */
 	void *dataptr = NULL; /* place to store file data */
